add optional keyword argument "keep_timestamps" to `method PriceHistory.history`
keep_timestamps defaults to false, so the behaviour of the method does
not change if the argument is ommited.

If keep_timestamps is True, the datetime of the OHLC data is added as an
extra column to the data frame. As far as I understand, the timestamps
are Unix times (in UTC).

For this to work, method `parse_quotes` in utils.py gets the same
optional keword argument.